+++ src/plugins/loaders.ts
@@ -39,4 +39,5 @@ export const loaders: Record<string, () => Promise<Source>> = {
   img: () => import('./img').then((m) => m.img),
   'img-color': () => import('./img-color').then((m) => m.imgColor),
   'video-color': () => import('./video-color').then((m) => m.videoColor),
+  truncated: () => import('./truncated').then((m) => m.truncated),
 }